feat(seed/index.js):  delete seed file, create seed directory and index file
